Improve composability for browser modules
